Split e2e tests into multiple files
